refactor: move sisu burst module
