+++ src/render/renderer.ts
@@ -2,7 +2,7 @@ import type { AxialCoord, PixelCoord } from '../hex/HexUtils.ts';
 import { hexDistance } from '../hex/HexUtils.ts';
 import type { LoadedAssets } from '../loader.ts';
 import type { Unit } from '../unit/index.ts';
-import { isSisuBurstActive } from '../sim/sisu.ts';
+import { isSisuBurstActive } from '../sisu/burst.ts';
 import type { Sauna } from '../sim/sauna.ts';
 import { HexMapRenderer } from './HexMapRenderer.ts';
 import { camera } from '../camera/autoFrame.ts';
